Found a better, more reliable solution for replicating the ball, and ridiciously simple to implement , compared to my previous solution based on interpolation/extrapolation . Ball is now synchronized between server & client !

+++ Source/PowerBall/SmoothPhysicsState.h
@@ -8,7 +8,7 @@ struct FSmoothPhysicsState
 	GENERATED_USTRUCT_BODY()
 
 	UPROPERTY()
-	float timestamp;
+	uint64 timestamp;
 	UPROPERTY()
 	FVector pos;
 	UPROPERTY()
